modificare din float in int

+++ src/Scadere.h
@@ -1,1 +1,1 @@
-double scadere(float x,float y);
+double scadere(int x,int y);
